backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
pytest>=8.0
pytest-django>=4.8
